add employee auth and others related data

+++ app/Http/Controllers/Dashboard/CompanyController.php
@@ -9,9 +9,11 @@ use App\Interfaces\CompanyRepositoryInterface;
 use App\Models\Company;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CompanyController extends BaseController
 {
+    use AuthorizesRequests;
     protected mixed $crudRepository;
 
     public function __construct(CompanyRepositoryInterface $pattern)
@@ -22,7 +24,7 @@ class CompanyController extends BaseController
     public function index()
     {
         try {
-
+            $this->authorize('viewAny', Company::class);
             $companies = CompanyResource::collection($this->crudRepository->all(
                 [],
                 [],
@@ -37,6 +39,8 @@ class CompanyController extends BaseController
     public function store(CompanyRequest $request)
     {
             try {
+
+                $this->authorize('create', Company::class);
                 $company = $this->crudRepository->create($request->validated());
                 
                 return new CompanyResource($company);
@@ -48,6 +52,11 @@ class CompanyController extends BaseController
     public function show(Company $company): ?\Illuminate\Http\JsonResponse
     {
         try {
+            if(Auth()->guard('employees')->check()) {
+              $company = $this->crudRepository->find(auth("employees")->user()->company_id);
+            }
+            $this->authorize('view', $company);
+            
             return JsonResponse::respondSuccess('Item Fetched Successfully', new CompanyResource($company));
         } catch (Exception $e) {
             return JsonResponse::respondError($e->getMessage());
@@ -57,16 +66,26 @@ class CompanyController extends BaseController
 
     public function update(CompanyRequest $request, Company $company)
     {
-        $this->crudRepository->update($request->validated(), $company->id);
+        try {
+            if(Auth()->guard('employees')->check()) {
+              $company = $this->crudRepository->find(auth("employees")->user()->company_id);
+            }
+            $this->authorize('update', $company);
+            $this->crudRepository->update($request->validated(), $company->id);
 
-     
-        return JsonResponse::respondSuccess(trans(JsonResponse::MSG_UPDATED_SUCCESSFULLY));
+            return JsonResponse::respondSuccess(trans(JsonResponse::MSG_UPDATED_SUCCESSFULLY));
+
+        } catch (\Throwable $th) {
+            return JsonResponse::respondError($th->getMessage());
+        }
+        
     }
 
 
     public function destroy(Request $request): ?\Illuminate\Http\JsonResponse
     {
         try {
+            $this->authorize('delete', Company::class);
             $this->crudRepository->deleteRecords('companies', $request['items']);
             return JsonResponse::respondSuccess(trans(JsonResponse::MSG_DELETED_SUCCESSFULLY));
         } catch (Exception $e) {
@@ -77,6 +96,7 @@ class CompanyController extends BaseController
     public function restore(Request $request): \Illuminate\Http\JsonResponse
     {
         try {
+            $this->authorize('restore', Company::class);
             $this->crudRepository->restoreItem(Company::class, $request['items']);
             return JsonResponse::respondSuccess(trans(JsonResponse::MSG_RESTORED_SUCCESSFULLY));
         } catch (Exception $e) {
@@ -90,6 +110,7 @@ class CompanyController extends BaseController
     public function forceDelete(Request $request): \Illuminate\Http\JsonResponse
     {
         try {
+            $this->authorize('forceDelete', Company::class);
             $this->crudRepository->deleteRecordsFinial(Company::class, $request['items']);
             return JsonResponse::respondSuccess(trans(JsonResponse::MSG_FORCE_DELETED_SUCCESSFULLY));
         } catch (Exception $e) {
